+++ cwingking.txt
@@ -1,1 +1,1 @@
-./Debug/src_windows_windows_graphics.c.o ./Debug/src_windows_windows_init.c.o ./Debug/src_md407_md407_asciidisplay.c.o ./Debug/src_graphics.c.o ./Debug/src_startup.c.o ./Debug/src_md407_md407_graphics.c.o ./Debug/src_snake_obsticalManager.c.o ./Debug/src_windows_windows_platform.c.o ./Debug/src_md407_md407_input.c.o ./Debug/src_splashScreen.c.o ./Debug/src_gameoverScreen.c.o ./Debug/src_windows_windows_delay.c.o ./Debug/src_rng.c.o ./Debug/src_windows_windows_input.c.o ./Debug/src_snake_roof.c.o ./Debug/src_snake_rope.c.o ./Debug/src_windows_windows_asciidisplay.c.o ./Debug/src_md407_md407_delay.c.o ./Debug/src_snake_obstacle.c.o ./Debug/src_snake_snake.c.o ./Debug/src_md407_md407_platform.c.o ./Debug/src_md407_md407_init.c.o   
+./Debug/src_splashscreen.c.o ./Debug/src_windows_windows_graphics.c.o ./Debug/src_windows_windows_init.c.o ./Debug/src_gameover_screen.c.o ./Debug/src_md407_md407_asciidisplay.c.o ./Debug/src_graphics.c.o ./Debug/src_startup.c.o ./Debug/src_md407_md407_graphics.c.o ./Debug/src_snake_obsticalManager.c.o ./Debug/src_windows_windows_platform.c.o ./Debug/src_md407_md407_input.c.o ./Debug/src_windows_windows_delay.c.o ./Debug/src_rng.c.o ./Debug/src_windows_windows_input.c.o ./Debug/src_snake_roof.c.o ./Debug/src_snake_rope.c.o ./Debug/src_windows_windows_asciidisplay.c.o ./Debug/src_md407_md407_delay.c.o ./Debug/src_snake_obstacle.c.o ./Debug/src_snake_snake.c.o ./Debug/src_md407_md407_platform.c.o ./Debug/src_md407_md407_init.c.o
