+++ makefile
@@ -2,8 +2,8 @@
 
 
 o_files = windowsout/startup.o windowsout/windows_graphics.o windowsout/windows_delay.o windowsout/windows_input.o windowsout/snake.o windowsout/windows_init.o \
-windowsout/windows_platform.o windowsout/obstacle.o windowsout/windows_asciidisplay.o windowsout/splashScreen.o \
-windowsout/gameoverScreen.o windowsout/roof.o windowsout/rope.o windowsout/obsticalManager.o windowsout/rng.o
+windowsout/windows_platform.o windowsout/obstacle.o windowsout/windows_asciidisplay.o windowsout/splashscreen.o \
+windowsout/gameover_screen.o windowsout/roof.o windowsout/rope.o windowsout/obstacle_manager.o windowsout/rng.o
 
 h_includes = -I src/snake -I src/windows -I src/windows/SDL/include -I src/windows/SDL/include/SDL2 -I src
 
@@ -45,10 +45,10 @@ windowsout/obstacle.o: src/snake/obstacle.c
 windowsout/windows_asciidisplay.o: src/windows/windows_asciidisplay.c
 	gcc -g -c $< $(define_flags) -o $@ $(h_includes)
 
-windowsout/splashScreen.o: src/splashScreen.c
+windowsout/splashscreen.o: src/splashscreen.c
 	gcc -g -c $< $(define_flags) -o $@ $(h_includes)
 
-windowsout/gameoverScreen.o: src/gameoverScreen.c
+windowsout/gameover_screen.o: src/gameover_screen.c
 	gcc -g -c $< $(define_flags) -o $@ $(h_includes)
 
 windowsout/roof.o: src/snake/roof.c
@@ -58,7 +58,7 @@ windowsout/roof.o: src/snake/roof.c
 windowsout/rope.o: src/snake/rope.c
 	gcc -g -c $< $(define_flags) -o $@ $(h_includes)
 
-windowsout/obsticalManager.o: src/snake/obsticalManager.c
+windowsout/obstacle_manager.o: src/snake/obstacle_manager.c
 	gcc -g -c $< $(define_flags) -o $@ $(h_includes)
 
 windowsout/rng.o: src/rng.c
