--- src/gameoverScreen.h
@@ -1,7 +1,0 @@
-#ifndef GAMEOVERSCREEN_H
-#define GAMEOVERSCREEN_H
-
-// Displays a gameover-screen
-void displayGameoverScreen(int score);
-
-#endif //GAMEOVERSCREEN
