+++ src/gameover_screen.h
@@ -1,0 +1,7 @@
+#ifndef GAMEOVER_SCREEN_H
+#define GAMEOVER_SCREEN_H
+
+// Displays a gameover-screen
+void displayGameoverScreen(int score);
+
+#endif //GAMEOVER_SCREEN
