+++ src/snake/obstacle_manager.c
