+++ src/startup.c
@@ -7,8 +7,8 @@
 #include "platform.h"
 #include "obstacle.h"
 #include "asciidisplay.h"
-#include "splashScreen.h"
-#include "gameoverScreen.h"
+#include "splashscreen.h"
+#include "gameover_screen.h"
 #include "roof.h"
 #include "rope.h"
 #include "obsticalManager.h"
